Audio FingerPrinting
FingerPrint class created to enable uploading and fingerprinting audio
files for detection.
GetDuration() of Song added to convert and calculate the duration of
fingerprinted audio.
GetAudioPath() and GetFullAudioPath() have been added to Song to return
the absolute path of audio file to be fingerprinted.
Duration of songs in list view updated to reflect actual duration.